done exercise 4.21*

+++ part 4/bloglist/controllers/blogs.js
@@ -10,7 +10,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body 
-  
+
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
   if (!decodedToken.id) {
@@ -38,8 +38,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blogToDelete = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token valid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if (blogToDelete.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).json({ 'error': 'Not authorized to delete this blog.' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
